Canonicalize special cases of `map` (5) move to own module
